more logging updates.

+++ src/Shapeshifter.WindowsDesktop/Infrastructure/Logging/FileLogStream.cs
@@ -14,7 +14,10 @@
 
         public FileLogStream()
         {
-            File.WriteAllText(logFileName, string.Empty);
+            if (!File.Exists(logFileName))
+            {
+                File.WriteAllText(logFileName, string.Empty);
+            }
         }
 
         public void WriteLine(string input)
